Add in checks for if theyre only admin or have any in progress requests

+++ src/registrar/views/portfolios.py
@@ -1,5 +1,5 @@
 import logging
-from django.http import HttpResponse, Http404
+from django.http import HttpResponse, Http404, JsonResponse
 from django.shortcuts import render
 from django.urls import reverse
 from django.contrib import messages
@@ -100,14 +100,26 @@ class PortfolioMemberView(PortfolioMemberPermissionView, View):
             },
         )
 
-class PortfolioMemberDeleteView(PortfolioMemberPermission, View):    
-    
+
+class PortfolioMemberDeleteView(PortfolioMemberPermission, View):
+
     def post(self, request, pk):
         """
         Find and delete the portfolio member using the provided primary key (pk).
         Redirect to a success page after deletion (or any other appropriate page).
         """
         portfolio_member_permission = get_object_or_404(UserPortfolioPermission, pk=pk)
+        member = portfolio_member_permission.user
+
+        active_requests_count = member.get_active_requests_count_in_portfolio(request)
+        print(f"Active requests count for member {member.id}: {active_requests_count}")
+
+        if active_requests_count > 0:
+            return JsonResponse({"error": "ERROR: Member has in-progress requests and cannot be removed."}, status=400)
+
+        # If they are the last manager of a domain
+        if member.is_only_admin_of_portfolio(portfolio_member_permission.portfolio):
+            return JsonResponse({"error": "ERROR: Member is the only admin."}, status=400)
 
         portfolio_member_permission.delete()
 
@@ -191,8 +203,8 @@ class PortfolioInvitedMemberView(PortfolioInvitedMemberPermissionView, View):
         )
 
 
-class PortfolioInvitedMemberDeleteView(PortfolioInvitedMemberPermission, View):    
-    
+class PortfolioInvitedMemberDeleteView(PortfolioInvitedMemberPermission, View):
+
     def post(self, request, pk):
         """
         Find and delete the portfolio invited member using the provided primary key (pk).
